refactor: redesign coursify cli using modular command structure
