+++ packages/coursify-cli/src/commands/index.js
@@ -1,0 +1,4 @@
+export { authCommands } from './auth.js';
+export { courseCommands } from './course.js';
+export { moduleCommands } from './module.js';
+export { sectionCommands } from './section.js';
